backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.3
motor==3.3.2
python-jose[cryptography]>=3.3.0 
passlib[bcrypt]>=1.7.4   
google-auth==2.27.0
fal-client==0.5.9
httpx==0.27.2
pytest==9.0.1
pytest-asyncio==1.3.0